+++ packages/vtable/examples/menu.ts
@@ -17,6 +17,10 @@ export const menus = [
       {
         path: 'debug',
         name: 'animation'
+      },
+      {
+        path: 'debug',
+        name: 'scroll'
       }
     ]
   },
@@ -520,6 +524,18 @@ export const menus = [
       {
         path: 'pivot-analysis',
         name: 'pivot-analysis-str'
+      },
+      {
+        path: 'pivot-analysis',
+        name: 'pivot-empty-tip'
+      },
+      {
+        path: 'pivot-analysis',
+        name: 'pivot-analysis-calculatedField'
+      },
+      {
+        path: 'pivot-analysis',
+        name: 'pivot-analysis-calculatedField2'
       }
     ]
   },